# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/appointments/{appointment}/payments": {
            "get": {
                "tags": [
                    "Payments"
                ],
                "summary": "Randevu ödemelerini listele",
                "operationId": "1380672dd5312ab437ed19b0d7a89ee1",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "description": "Randevu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ödemeler başarıyla listelendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payments": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "format": "float"
                                                    },
                                                    "note": {
                                                        "type": "string",
                                                        "nullable": true
                                                    },
                                                    "customer_id": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total_paid": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "remaining": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "is_paid": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Randevuya yeni ödeme ekle",
                "operationId": "1ba41e28abce5c390bad00240b175c4f",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "description": "Randevu ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "note": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ödeme başarıyla kaydedildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "payment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float"
                                                },
                                                "note": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "customer_id": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Geçersiz ödeme",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}